fix(core): use focusin event to track list-item focus
