+++ apps/docs/src/app/core/component-docs/list/examples/list-navigation-example/list-navigation-example.component.html
@@ -10,8 +10,8 @@
             <span fd-list-title> List item 2 </span>
         </a>
     </li>
-    <li fd-list-item>
-        <a fd-list-link [selected]="true">
+    <li fd-list-item [selected]="true">
+        <a fd-list-link>
             <span fd-list-title> List item 3 </span>
             <i fd-list-icon glyph="map"></i>
         </a>
