+++ libs/core/src/lib/list/list-focus-item.model.ts
@@ -1,11 +1,25 @@
-import { Directive, ElementRef, EventEmitter, HostListener } from '@angular/core';
+import { coerceNumberProperty } from '@angular/cdk/coercion';
+import { Directive, ElementRef, EventEmitter, HostBinding, HostListener, Input } from '@angular/core';
 import { KeyboardSupportItemInterface } from '@fundamental-ngx/core/utils';
 import { Subject } from 'rxjs';
 
 @Directive()
 export abstract class ListFocusItem implements KeyboardSupportItemInterface {
+    /** tab index attribute */
+    @Input()
+    @HostBinding('attr.tabindex')
+    get tabindex(): number {
+        if (this._isFirstItem && isNaN(this._tabIndex)) {
+            return 0;
+        }
+        return this._tabIndex ?? -1;
+    }
+    set tabindex(value: number) {
+        this._tabIndex = coerceNumberProperty(value, -1);
+    }
+
     /** @hidden */
-    readonly _focused$ = new Subject<void>();
+    readonly _focused$ = new Subject<{ focusedWithin: boolean }>();
 
     /** @hidden */
     readonly _clicked$ = new Subject<MouseEvent>();
@@ -13,13 +27,18 @@ export abstract class ListFocusItem implements KeyboardSupportItemInterface {
     /** @hidden */
     protected _isFirstItem = false;
 
+    /** @hidden */
+    protected _tabIndex: number | undefined;
+
     /** @hidden Implementation of KeyboardSupportItemInterface */
     keyDown: EventEmitter<KeyboardEvent> = new EventEmitter<KeyboardEvent>();
 
     /** @hidden */
-    @HostListener('focus')
-    protected onFocus(): void {
-        this._focused$.next();
+    @HostListener('focusin', ['$event'])
+    protected onFocus(event: FocusEvent): void {
+        this._focused$.next({
+            focusedWithin: event.target !== this.elementRef?.nativeElement
+        });
     }
 
     constructor(readonly elementRef: ElementRef) {}
