+++ libs/core/src/lib/list/list-item/list-item.component.html
@@ -11,6 +11,6 @@
     {{ counter | number: '1.0-2' }}
 </span>
 
-<span class="fd-screenreader-only" aria-hidden="true" [id]="uniqueId + '-screenReader-content'">{{
+<span class="fd-screenreader-only" aria-hidden="true" [id]="_uniqueId + '-screenReader-content'">{{
     screenReaderContent
 }}</span>
