+++ libs/core/src/lib/list/list-item/list-item.component.scss
@@ -5,8 +5,6 @@
     height: 1px;
     width: 1px;
     overflow: hidden;
-    /* IE 7+ only support clip without commas */
-    clip: rect(1px 1px 1px 1px);
     /* All other browsers */
     clip: rect(1px, 1px, 1px, 1px);
 }
